requirements.txt
schedule
python-dotenv
sentry-sdk
